requirements.txt
absl-py==0.13.0
aioice==0.7.5
aiortc==1.2.0
altair==4.1.0
altgraph==0.17
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
autopy==4.0.0
av==8.0.3
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.10
crc32c==2.2.post0
cryptography==3.4.7
cvzone==1.3.7
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
dnspython==2.1.0
Eel==0.12.4
entrypoints==0.3
enum34==1.1.10
face-recognition-models==0.3.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gevent==21.1.2
gevent-websocket==0.10.1
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
ipykernel==6.0.0
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mediapipe==0.8.5
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.11.0
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pefile==2021.5.24
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycaw==20181226
pycparser==2.20
pydeck==0.6.2
pydub==0.25.1
pyee==8.1.0
Pygments==2.9.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pylibsrtp==0.6.8
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
pyvirtualcam==0.8.0
pywin32==301
pywin32-ctypes==0.2.0
pywinpty==1.1.3
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Send2Trash==1.7.1
six==1.15.0
smmap==4.0.0
streamlit==0.83.0
streamlit-webrtc==0.23.3
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.5
validators==0.18.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
